pyproject.toml
[project]
name = "python-mono-repo-starter-kit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "rich",
    "typer-slim[standard]",
    "uvicorn[standard]>=0.40.0",
    "core",
    "shared",
    "python-dotenv>=1.2.1",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
]

[project.scripts]
#Microservice necessary components
api-gateway = "api_gateway.main:app"
config-server = "config_server.main:app"
service-discovery = "service_discovery.main:app"

#Microservice app base
user = "user.main:app"
item = "item.main:app"
item-management = "item_management.main:app"
product = "product.main:app"
review-and-ratings = "review_and_ratings.main:app"
inventory = "inventory.main:app"
recommendations = "recommendations.main:app"
offers = "offers.main:app"
cart = "cart.main:app"
order = "order.main:app"
archival = "archival.main:app"
notification = "notification.main:app"
serviceability = "serviceability.main:app"
payment = "payment.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [ 
    "services/api-service",
    "services/api-gateway",
    "services/config-server",
    "services/service-discovery",
    "services/user",
    "services/item",
    "services/item-management",
    "services/product",
    "services/review-and-ratings",
    "services/inventory",
    "services/recommendations",
    "services/offers",
    "services/cart",
    "services/order",
    "services/archival",
    "services/notification",
    "services/serviceability",
    "services/payment",
    "libs/shared",
    "packages/core",]

[dependency-groups]
nox = [
    "nox-uv", 
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-randomly",
]

type_check = [
    "mypy",
]

lint = [
    "ruff",
]

docs = [
    "mkdocs-material",
    "mkdocs-htmlproofer-plugin",
    "mkdocstrings[python]",
    "mkdocs-api-autonav",
]

licenses = [
    "pip-licenses-cli",
]

debugging_tool = [
    "pysnooper",
]

[tool.setuptools.packages.find]
where = ["services", "pacakges", "libs"]

[tool.uv.sources]
core = { workspace = true }
shared = { workspace = true, editable = true }

[tool.uv.workspace]
members = [
    "services/api-service",
    "services/api-gateway",
    "services/config-server",
    "services/service-discovery",
    "services/user",
    "services/item",
    "services/item-management",
    "services/product",
    "services/review-and-ratings",
    "services/inventory",
    "services/recommendations",
    "services/offers",
    "services/cart",
    "services/order",
    "services/archival",
    "services/notification",
    "services/serviceability",
    "services/payment",
    "libs/shared",
    "packages/core",
]

[tool.mypy]
ignore_missing_imports = true
strict = true
enable_error_code = ["deprecated", "exhaustive-match", "explicit-override"]

[tool.ruff]
src = ["services", "pacakges", "libs"]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4
target-version = "py314"

[tool.ruff.lint]
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "I", # isort
    "N", # pep8-naming
    "UP", # pyupgrade
    "RUF", # ruff
    "B", # flake8-bugbear
    "C4", # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "PT", # flake-pytest-style
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
ignore = []
fixable = ["ALL"]
extend-ignore = [
    "RUF005",
    "RUF012",
]
unfixable = [
    "F401",
]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
